+++ packages/mui-material/src/Select/SelectInput.js
@@ -19,6 +19,11 @@ import useForkRef from '../utils/useForkRef';
 import useControlled from '../utils/useControlled';
 import selectClasses, { getSelectUtilityClasses } from './selectClasses';
 import { areEqualValues, isEmpty, getOpenInteractionType } from './utils';
+import {
+  canCycleRepeatedCharacter,
+  getMatchingOptionIndex,
+  getTypeaheadOptions,
+} from './utils/closedTypeahead';
 import { SelectFocusSourceProvider } from './utils/SelectFocusSourceContext';
 
 const OPENING_MOUSE_UP_BOUNDARY_OFFSET = 2;
@@ -30,85 +35,6 @@ const SPACE = ' ';
 const ARROW_UP = 'ArrowUp';
 const ARROW_DOWN = 'ArrowDown';
 const ENTER = 'Enter';
-
-function hasOwnValueProp(child) {
-  return Object.prototype.hasOwnProperty.call(child.props, 'value');
-}
-
-function getTextFromReactNode(node) {
-  if (typeof node === 'string' || typeof node === 'number') {
-    return String(node);
-  }
-
-  let text = '';
-
-  React.Children.forEach(node, (child) => {
-    if (typeof child === 'string' || typeof child === 'number') {
-      text += String(child);
-    } else if (React.isValidElement(child)) {
-      text += getTextFromReactNode(child.props.children);
-    }
-  });
-
-  return text;
-}
-
-function getMatchingOptionIndex(options, search, startIndex = 0) {
-  if (options.length === 0) {
-    return -1;
-  }
-
-  const normalizedStartIndex = ((startIndex % options.length) + options.length) % options.length;
-
-  for (let offset = 0; offset < options.length; offset += 1) {
-    const index = (normalizedStartIndex + offset) % options.length;
-
-    if (options[index].label.startsWith(search)) {
-      return index;
-    }
-  }
-
-  return -1;
-}
-
-function canCycleRepeatedCharacter(options, key) {
-  return !options.some((option) => option.label[0] === key && option.label[1] === key);
-}
-
-function getTypeaheadOptions(childrenArray, value) {
-  const options = [];
-  let selectedIndex = -1;
-
-  for (let index = 0; index < childrenArray.length; index += 1) {
-    const child = childrenArray[index];
-
-    if (!React.isValidElement(child) || !hasOwnValueProp(child) || child.props.disabled) {
-      continue;
-    }
-
-    // Closed typeahead cannot exclude CSS-hidden text because no option DOM is mounted.
-    const label = getTextFromReactNode(child.props.children).trim().toLowerCase();
-
-    if (label === '') {
-      continue;
-    }
-
-    if (selectedIndex === -1 && areEqualValues(value, child.props.value)) {
-      selectedIndex = options.length;
-    }
-
-    options.push({
-      child,
-      label,
-      value: child.props.value,
-    });
-  }
-
-  return {
-    options,
-    selectedIndex,
-  };
-}
 
 /**
  * Returns true when a native mouse event should be treated as happening inside
